+++ modules/http-fauth/src/stage.rs
@@ -118,7 +118,11 @@ impl ForwardedAuthenticationStage {
             .query()
             .map(|q| format!("?{}", q))
             .unwrap_or_default();
-        let auth_uri = format!("{}{}{}", config.backend_url, path, query);
+        let auth_uri = if config.no_append_uri {
+            config.backend_url.clone()
+        } else {
+            format!("{}{}{}", config.backend_url, path, query)
+        };
 
         let mut auth_request = Request::builder()
             .uri(auth_uri)